MiniBatchInference.jl

MiniBatchInference.ResultMLE — Type

Container for ouputs of MLE.

Notes

res = ResultMLE() has all fields empty but res.minloss which is set to Inf.

MiniBatchInference.AIC — Method

AIC(RSS, k, m)

Calculate AIC of a model given its RSS, k its number of parameters, and m the number of observations.

MiniBatchInference.AIC — Method

AIC(res, data_set, Σ)

Computes the AIC of res given the observational noise variance covariance matrix Σ.

MiniBatchInference.AIC — Method

AIC(logl, nparams)

Computes the AIC given loglikelihood logl and number of parameters nparams.

MiniBatchInference.AICc — Method

AICc(aic, k, m)

Calculate AIC corrected of a model given its aic, k its number of parameters, and m the number of observations (if d variables and T time steps, m = d * T). The formula is taken from Mangan2017.

MiniBatchInference.AICc_TREE — Method

AICc_TREE(RSS, k, m)

Calculate aic of a model given its RSS, k its number of parameters, and m the number of observations. The formula is taken from TREE article.

MiniBatchInference.FIM_strouwen — Method

FIM_strouwen(predict, θ, Σ)

Returns the FIM matrix associated to predict and evaluated at θ taken from https://arnostrouwen.com/post/dynamic-experimental-design/

predict is a function that takes θ as a parameter and returns an array

with dim=1 corresponding to state variables and

dim = 2 corresponding to the time steps

θ the parameter vector where to evaluate the FIM

Σ is the variance-covariance matrix of the observation errors

MiniBatchInference.FIM_yazdani — Method

FIM_yazdani(dudt, u0_true, tspan, p_true, Σ)

Returns the FIM matrix associated to problem defined by prob = ODEProblem(dudt, u0_true, tspan, p_true). Σ is the variance-covariance matrix of the observation errors

MiniBatchInference.RSS — Function

RSS(res, data_set)
RSS(res, data_set, fn)

Computes the RSS of res given data_set.

Argument:

fn is set to identity but can be e.g. log if lognormal loss used.

MiniBatchInference._loss_multiple_shoot_init — Method

_loss_multiple_shoot_init(data, pred, ic_term)

default loss function for minibatch_MLE.

MiniBatchInference.divisors — Method

divisors(n)

Returns all divisors of n, sorted.

MiniBatchInference.get_u0s — Method

get_u0s(res)

Returns initial condition vector estimated[u_0_1, ..., u_0_n] , where n corresponds to the number of chunks. In the case of independent time series, returns [[u_0_TS1_1, ..., u_0_TS1_n],...,[u_0_TS1_1,...]], matching the format of res.pred.

MiniBatchInference.iterative_minibatch_MLE — Method

iterative_minibatch_MLE(; group_sizes, optimizers_array, threshold, kwargs...)

Performs a iterative minibatch MLE, iterating over group_sizes. Stops the iteration when loss function increases between two iterations.

Returns an array with all ResultMLE obtained during the iteration. For kwargs, see minibatch_MLE.

Note

for now, does not support independent time series (minibatch_ML_indep_TS).

at every iteration, initial conditions are initialised given the predition of previous iterations

Specific arguments

group_sizes : array of group sizes to test

optimizers_array: optimizersarray[i] is an array of optimizers for the trainging processe of `groupsizes[i`

MiniBatchInference.loglikelihood — Method

loglikelihood(res, data_set, Σ; distrib, fn)

Computes the loglikelihood of res given the observational noise variance covariance matrix Σ.

Options

By default, normal observational noise is assumed, but lognormal observational noise can be chosen by setting distrib=MvLogNormal, fn=log

MiniBatchInference.minibatch_MLE — Method

minibatch_MLE(; group_size, kwargs...)

Maximum likelihood estimation with minibatching. Loops through ADAM and BFGS. Returns minloss, p_trained, ranges, losses, θs.

arguments

p_init : initial guess for parameters of prob

group_size : size of segments

data_set : data

prob : ODE problem for the state variables.

tsteps : corresponding to data

alg : ODE solver

sensealg : sensitivity solver

optional

loss_fn : the loss function, that takes as arguments loss_fn(data, pred, ic_term) where data is the training data and pred corresponds to the predicted state variables. loss_fn must transform the pred into the observables, with a function h that maps the state variables to the observables. By default, h is taken as the identity.

u0_init : if not provided, we initialise from data_set

loss_fn : loss function with arguments loss_fn(data, pred, ic_term)

λ : dictionary with learning rates. Dict("ADAM" => 0.01, "BFGS" => 0.01)

maxiters : dictionary with maximum iterations. Dict("ADAM" => 2000, "BFGS" => 1000),

continuity_term : weight on continuity conditions

ic_term : weight on initial conditions

verbose : displaying loss

info_per_its = 50,

plotting : plotting convergence loss

info_per_its = 50,

cb : call back function. Must be of the form cb(θs, p_trained, losses, pred, ranges)

p_true : true params

p_labs : labels of the true parameters

threshold : default to 1e-6

MiniBatchInference.minibatch_ML_indep_TS — Method

minibatch_ML_indep_TS(; group_size, data_set, tsteps, kwargs...)

Similar to minibatch_MLE but for independent time series, where data_set is a vector containing the independent arrays corresponding to the time series, and tsteps is a vector where each entry contains the time steps of the corresponding time series.

MiniBatchInference.minibatch_loss — Method

minibatch_loss(θ, ode_data, tsteps, prob, loss_function, continuity_loss, solver, ranges; continuity_term, kwargs...)

Returns a tuple (loss, pred) obtained from minibatching of the time series ode_data into segments with time steps given by tsteps[ranges[i]]. The initial conditions are assumed free parameters for each segments.

Arguments:

θ: [u0,p] where p corresponds to the parameters of ode function.

ode_data: Original Data to be modelled.

tsteps: Timesteps on which ode_data was calculated.

prob: ODE problem that the Neural Network attempts to solve.

loss_function: Any arbitrary function to calculate loss.

continuity_loss: Function that takes states $\hat{u}_{end}$ of group $k$ and

$u_{0}$ of group $k+1$ as input and calculates prediction continuity loss between them. If no custom continuity_loss is specified, sum(abs, û_end - u_0) is used.

solver: ODE Solver algorithm.

ranges: Vector containg range for each segment.

continuity_term: Weight term to ensure continuity of predictions throughout different groups.

kwargs: Additional arguments splatted to the ODE solver. Refer to the

Local Sensitivity Analysis and Common Solver Arguments documentation for more details.

Note:

The parameter 'continuity_term' should be a relatively big number to enforce a large penalty whenever the last point of any group doesn't coincide with the first point of next group.

MiniBatchInference.moments! — Method

moments!(xx, x)

reurns the moments of x as a vector stored in xx

Args

xx : the covariate vector

x : the state variable vector

MiniBatchInference.moments — Method

moments(x)

reurns the moments of x as a vector

Args

xx : the covariate vector

x : the state variable vector